Remove extra leading space on tag names in ComicInfo
Komga might be stripping these out but on LANraragi this causes an issue.  Visually it removes it from the links the labels but it fails to find anything unless you go up and add that extra space into the search.  I've removed it and checked both Komga and LANraragi.  Komga still works normally and LANraragi now searches proper.