Relocate web archive pull
Moved the pull so it isn't part of the main loop.  I did duplicate the download image function with the only part removing the server loop and replacing the image download part with the one that rewrites and attempts to download the web archive version.  Previous version of the code tried to download from the archvie 5 times (once for each attempt pass) and I didn't realize this at first.  This one only does it once and only after the 5 tries have been tried with the regular servers.  Ideally I'd like to reduce the code duplicate (or eliminate outright) and have it abort the archive attempt at the first error.  At least it is better about it now.